requirements.txt
altair==4.2.2
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
aniso8601==9.0.1
anyio==3.7.0
appnope==0.1.3
APScheduler==3.10.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.3
configobj==5.0.8
configparser==5.3.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-design-kit==0.0.1
dash-html-components==2.0.0
dash-leaflet==0.1.28
dash-table==5.0.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
Django==4.1
entrypoints==0.4
etelemetry==0.3.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
Fiona==1.9.4.post1
fitz==0.0.1.dev2
Flask==2.2.5
Flask-RESTful==0.3.10
folium==0.14.1.dev59+g087d73b
fonttools==4.40.0
fqdn==1.5.1
future==0.18.3
geobuf==1.1.1
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.4.1
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
looseversion==1.2.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
moviepy==1.0.3
mpmath==1.3.0
mysql-connector-python==8.0.33
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
notebook_shim==0.2.3
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
overrides==7.3.1
OWSLib==0.29.2
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.2.0
plotly==5.14.1
portpicker==1.5.2
proglog==0.1.10
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
prov==2.0.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.15.0
Pympler==1.0.1
PyMuPDF==1.22.3
PyMySQL==1.0.3
pyparsing==3.1.0
PyPDF2==1.26.0
pyproj==3.6.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.5
PyYAML==6.0
pyzmq==25.0.2
rdflib==6.3.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
setuptools==65.6.3
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.3
stack-data==0.6.2
streamlit==1.25.0
streamlit-folium==0.13.0
streamlit-option-menu==0.3.6
streamlit-searchbox==0.1.3
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
vega-datasets==0.9.0
Wand==0.6.11
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==4.0.7
wsproto==1.2.0
zipp==3.15.0
